# fixed_pipeline_state: Pack structure, use memcmp and CityHash on it