… src/video_core/renderer_vulkan/fixed_pipeline_state.h
… src/video_core/renderer_vulkan/vk_pipeline_cache.cpp
… src/video_core/renderer_vulkan/vk_rasterizer.cpp