Compare the whole struct with std::memcmp and hash with CityHash. Using CityHash instead of a naive hash should reduce the number of collisions. Improve used type traits to ensure this operation is safe.

With these changes the improvements to the hashable pipeline state are:
Optimized structure
```
Hash:            89 ns
Equal:          103 ns
Construction*:  164 ns
Struct size:    384 bytes
```
Original structure
```
Hash:           148 ns
Equal:          174 ns
Construction*:  281 ns
Size:          1384 bytes
```
* Attribute state initialization is not measured

These measures are averages taken with `std::chrono::high_resolution_clock` on MSVC shipped on Visual Studio 16.6.0 Preview 2.1.